Add reflection to Engine::Object
